feat: full trading system — The Equation
Spectral alpha signal discovery -> regime rotation -> Mag 7 momentum
trading with leading indicators and defensive asset rotation.

- 87 features: spectral, price, macro, astro, weather
- Signal discovery: SPY/TLT, HYG, ARKK, VIX lead tech by 2-5 days
- Regime classifier: NORMAL/RISK_ON/FEAR/CRISIS/INFLATION/RECESSION
- Mag 7 momentum-weighted allocation with NVDA dip-buying
- Singularity override: when 6/7 tech stocks ripping, go all-in
- Defensive rotation: energy/gold/bonds when tech isn't the play
- Live trading via TastyTrade websocket streaming
- Multi-instance runner for A/B testing 10 variants on Raspberry Pi
- Backtest: $100K -> $355K in 4 years, Sharpe 1.46, MaxDD 19%

+++ wavegpt/spectral_linear.py
@@ -140,7 +140,7 @@ class SpectralLinear(nn.Module):
     def to_linear(self) -> nn.Linear:
         """Reconstruct a standard nn.Linear from current spectral params."""
         with torch.no_grad():
-            spectrum = self.get_spectrum()
+            spectrum = self.get_spectrum().to(self.U.dtype)
             W = (self.U * spectrum.unsqueeze(0)) @ self.V.t()
         linear = nn.Linear(self.in_dim, self.out_dim, bias=self.bias is not None)
         linear.weight.data = W
